refactor; split write and read for queries
