+++ cognite/neat/_graph/queries/__init__.py
@@ -1,3 +1,3 @@
-from ._base import Queries
+from ._queries import Queries
 
 __all__ = ["Queries"]
