+++ cognite/neat/_session/_set.py
@@ -76,8 +76,8 @@ class SetInstances:
             neat.set.instances.replace_type("Asset", "assetCategory")
             ```
         """
-        type_uri = self._state.instances.store.queries.type_uri(current_type)
-        property_uri = self._state.instances.store.queries.property_uri(property_type)
+        type_uri = self._state.instances.store.queries.read.type_uri(current_type)
+        property_uri = self._state.instances.store.queries.read.property_uri(property_type)
 
         if not type_uri:
             raise NeatValueError(f"Type {current_type} does not exist in the graph.")
@@ -93,7 +93,7 @@ class SetInstances:
                 f"{property_type} has multiple ids found in the graph: {humanize_collection(property_uri)}."
             )
 
-        if not self._state.instances.store.queries.type_with_property(type_uri[0], property_uri[0]):
+        if not self._state.instances.store.queries.read.type_with_property(type_uri[0], property_uri[0]):
             raise NeatValueError(f"Property {property_type} is not defined for type {current_type}.")
 
         self._state.instances.store.transform(SetType(type_uri[0], property_uri[0], drop_property))
